+++ src/main/resources/mybatis/mappers/SalaryPaymentMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.woosan.hr_system.salary.dao.SalaryPaymentDAO">
+<mapper namespace="salaryPayment">
     <resultMap id="SalaryPaymentResultMap" type="com.woosan.hr_system.salary.model.SalaryPayment">
         <id property="paymentId" column="payment_id" />
         <result property="salaryId" column="salary_id" />
